backend/voice/requirements_ml.txt
# ML Training Requirements for JARVIS Voice System
# Install with: pip install -r requirements_ml.txt

# Core ML libraries
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
noisereduce>=3.0.0

# Deep learning (required for hybrid STT)
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0  # For Wav2Vec 2.0

# STT Engines
vosk>=0.3.45  # Lightweight offline STT
openai-whisper>=20231117  # Local Whisper support

# HTTP client for GCP Whisper
aiohttp>=3.9.0

# System resources monitoring
psutil>=5.9.0

# Data handling
joblib>=1.3.0
pandas>=2.0.0

# Additional voice dependencies
webrtcvad>=2.0.10
pyaudio>=0.2.11

# Existing voice requirements
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pygame>=2.5.0

# Anthropic API
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
